Product Admin - securing the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, get_object_or_404, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -74,8 +75,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store"""
+
+    # Check if user is superuser
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry Only store owners are allowe!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +102,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store"""
+
+    # Check if user is superuser
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry Only store owners are allowe!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)  # prefilling the form by getting product
     #  Create a post handler for update button
     if request.method == 'POST':
@@ -119,8 +134,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a product from the store"""
+
+    # Check if user is superuser
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry Only store owners are allowe!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
